+++ OPAL/ProjectDependencies.mmd
@@ -42,8 +42,7 @@ flowchart BT
     ide --> si
     ide --> br
 
-    ifds --> si
-    ifds --> br
+    ifds --> ide
 
     ai --> br
 
@@ -52,7 +51,6 @@ flowchart BT
     de --> ai
 
     tac --> ifds
-    tac --> ide
     tac --> ai
 
     apk --> tac
