--- OPAL/tac/src/main/scala/org/opalj/tac/fpcf/analyses/ifds/JavaForwardICFG.scala
@@ -1,18 +1,0 @@
-/* BSD 2-Clause License - see OPAL/LICENSE for details. */
-package org.opalj
-package tac
-package fpcf
-package analyses
-package ifds
-
-import org.opalj.br.analyses.SomeProject
-
-/**
- * An ICFG for a Java IFDS forward analysis.
- *
- * @param project the project to which the ICFG belongs.
- *
- * @author Nicolas Gross
- */
-class JavaForwardICFG(project: SomeProject) extends org.opalj.tac.fpcf.analyses.ide.solver.JavaForwardICFG(project)
-    with JavaICFG {}
